[기능추가]/Calendar #30 일별 수입, 지출 합을 구하는 함수 util 분리
